requirements.txt
pygame == 1.9.6
neat-python ==  0.92